Added references to new elements for C-grid fluxes
  Added references to the new elements Ice%flux_uv_stagger, Ocean%stagger,
ice_ocean_boundary%wind_stagger, and ocean_ice_boundary%stagger, to enable the
passing of c-grid velocities and fluxes.  These elements need to be added for
other components to be able to compile, and should have values of BGRID_NE if
the previous answers are to be reproduced.  Also added code initializing the
various elements of the ice_ocean_boundary type to 0.  This code will not
compile unless these elements have been added to the appropriate types defined
in the ocean and sea-ice component models.