Initialize ocean_ice_boundary%stagger in AMIP mode.
If running in AMIP mode Ocean%stagger is not initialized (due to ocean pes not being valid)
SIS2 code checks for value of oib%stagger and fails as it doesn't recognize the value.

Fix here is to initialize OIB%stagger to be the same as the stagger used in the ice model.