Use the correct ice domains for I/O.
  There are several places where I/O is done, either to read_data or to registar
a field for restarts, where Ice%slow_domain_NH needed to be used in place of
Ice%domain in order to be able to use different layouts for the fast and slow
ice processes.  This commit also sets Verona_coupler=.false. in the call to
ice_model_init. All answers are bitwise identical and reproduce the solutions
with the same fast and slow layouts.