Removed dangerous Ice_Ocean_Flux module variables
  Eliminated the module variables nxc_ocn, nyc_ice, etc., from
ice_ocean_flux_exchange.  These had been used to store array sizes, and
represented a deplorably bad example of how not to write reliable
quasi-object-oriented code, as there is no guarantee that the routines in this
module would always be called with components with the same array sizes or
domain decompositions.  Also changed Ice%PE to Ice%slow_ice_PE or added a
logical qualifier to calls to mpp_set_current_pelist.  All answers are bitwise
identical in test cases.