Corrected PElist calls in coupler_main
  Corrected various calls to mpp_set_current_pelist for the case where the slow
sea-ice is run on the ocean PEs and the fast sea-ice is run on the atmospheric
PEs.  This configuration also required that several sea-ice code clocks be split
up to reflect the different PE lists they are timing.  The SIS2-based coupled
model now runs in this mode, provided that the SIS2 code is using commit
NOAA-GFDL/SIS2@95aca89e519ee99ef17fad2b91fc797fbaf81002 or later, but the
answers are still somewhat different (for as yet unknown reasons) from what is
obtained when the fast and slow ice are both run on the atmospheric PEs.  In
that previous mode, all answers are bitwise identical.