Fix for tracer flux derivatives over the ocean.
Previously, model used ex_dfdtr_surf(:,isphum) as a derivative of all
tracer fluxes w.r.t. the surface concentrations. This value is used, for
example, to ensure that the implicitly-calculated fluxes of tracers that
exchange with only one surface type (land or ocean) are zero over the
other surface type.

However, ex_dfdtr_surf(:,isphum) is manipulated in the surface_flux
subroutine: it is hard-set to zero over the ocean points, so it is not
appropriate to use for other tracers.

Since flux = rho*Cd*|v|*(q_surf-q_atm) we can simply use negative
dfdtr_atm for the dfdtr_surf derivative. This is what this fix does.

This fix is only changes anything for a very specific case, it should
not change answers in any realistic configurations of the models.