Added concurrent_ice coupling option
  Added a new namelist variable, concurrent_ice, to enable a new coupling mode.
In this mode, the fast sea ice is run on the atmospheric PEs, but the slow ice
is run with the ocean, taking the fast ice fluxes from the previous timestep and
redoing the shortwave and diffusive heating with the new ice state.  This mode
is concurrent between the atmosphere and ocean, and more efficient in some
configurations than using lagged fluxes.  By default, this mode is not used and
all solutions are bitwise identical.  This change requires the use of SIS2
commit NOAA-GFDL/SIS2@71149b80c9d9347524edef6e3fde8406c1e441e6 or later and the
corresponding changes to SIS1.