+Send three _berg fields from the ice to the ocean
  Three new fields to be passed from the ice to the ocean model: ustar_berg,
area_berg, mass_berg, as requested by @SternAlon. The fields are allocated in
the ice_ocean_boundary type if they are associated in the sea ice model. This
means that the code will not compile when used with older versions of SIS2 or
SIS, but all answers are bitwise identical for any cases that compile.