+Corrected legacy land interfaces
  Added cpp macros to restrict the land interface used by
atm_land_ice_flux_exchange to those that were available in the verona release.
After preprocessing, this code is unchanged if _USE_LEGACY_LAND_ is undefined,
but if it is defined and the verona land code is being used, this code would
not even compile.  All answers in MOM6-examples test cases using warsaw code
for everything but LM3 and verona code for LM3 are bitwise identical to pure
verona code.