Remove _USE_LEGACY_LAND_ ifndef around send_data call
A #ifndef _USE_LEGACY_LAND_ was added around a send_data call, and added a send_tile_average_data call that was not needed to allow a legacy land source version to run with the updated warsaw coupler.

Also added protection to other send_tile_data and send_tile_averaged_data calls to not call the routine if the id to send is less than 0.