Travis CI testing
This include a script that will build and run the full coupled null model.  This test will use the null component repositories on github, and a data file on the GFDL FTP server to run the tests on a travis CI image.  This also builds a simple coupler executable, but a run test is still not available.

Fixes #11