Remove archaic calls working on Ice%ocean_fluxes_top
  Removed archaic calls creating and setting up Ice%ocean_fluxes_top from the
coupler code.  This structure should have always been private to ice models, and
is not even used by any ice models that are still in active development or use,
but until now this wasteful, unused element had to be retained because of these
misplaced calls.  For any older ice models, the fact that the FMS coupler_types
code supports a copy from a 2d coupler type to a 3d coupler type means that this
could be set up inside of the ice model if it is still needed.  In the
MOM6-examples test suite, all answers are bitwise identical with these changes.