Add cheat mode
Add "cheat mode", which can be enabled via the `CHEAT_MODE_DIR` CPP
definition. When enabled, `coupler_init` will check for the existence of
a tarball corresponding to the current segment in the directory
specified by `CHEAT_MODE_DIR`. If this tarball already exists, the main
loop is skipped and the root PE extracts the tarball before the program
exits (after `fms_end` is called). If the tarball does not exist, the
main loop runs normally and the root PE creates the tarball before the
program exits (after `fms_end` is called).